+++ apps/modernjs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @module-federation/modernjs
 
+## 0.1.56
+
+### Patch Changes
+
+- Updated dependencies [6b02145]
+- Updated dependencies [22a3b83]
+  - @module-federation/enhanced@0.6.10
+
 ## 0.1.55
 
 ### Patch Changes